feature #2592
git-svn-id: svn+ssh://svn.amanzi.org/var/lib/svn/trunk/awe@5465 0edadee9-8154-4c59-a661-b5f45207fcbd

+++ org.amanzi.awe.views.reuse/src/org/amanzi/awe/views/reuse/views/Column.java
@@ -154,7 +154,7 @@ public class Column implements Comparable<Column> {
      *
      * @param countNode the new value
      */
-    public void setValue(int countNode) {
+    public void setValue(double countNode) {
         if (node != null) {
             node.setProperty(INeoConstants.PROPERTY_VALUE_NAME, countNode);
         }
